requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
langfuse>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional
pylint>=3.0.0
